refactored payouts, container, and photos
