+++ helpers/photos.helper.js
